// student/src/common/types/Student.ts
export default class Student {
  id: number = -1;
  registerNumber: string = "";
  name: string = "";
  college: string = "";
  course: string = "";
  sslcPercentage: number = 0;
  hscPercentage: number = 0;
  ugCgpa: number = 0;
  pgCgpa: number = 0;
  currentArrears: number = 0;
  historyOfArrears: number = 0;
  placedCount: number = 0;
  resume: string | null = null;
}